+++ R/fit.R
@@ -149,6 +149,7 @@
 #' @importFrom insight get_response get_data
 #' @importFrom cv GetResponse cv
 #' @importFrom sf st_relate st_coordinates st_centroid st_within st_crs
+#' @importFrom cli cli_abort cli_warn cli_inform cli_alert_success cli_alert_danger cli_alert_info
 #'
 #' @seealso Details section of [make_dsem_ram()] for a summary of the math involved with constructing the DSEM, and \doi{10.1111/2041-210X.14289} for more background on math and inference
 #' @seealso \doi{10.48550/arXiv.2401.10193} for more details on how GAM, SEM, and DSEM components are combined from a statistical and software-user perspective
@@ -1111,7 +1112,8 @@ function( formula,
     weights = weights,
     packageVersion = packageVersion("tinyVAST"),
     development = development,
-    distributions = distributions
+    distributions = distributions,
+    convergence_diagnostics = get_convergence_diagnostics(sdrep, obj)
   )
   out = list(
     data = data,