+++ R/methods.R
@@ -25,6 +25,13 @@ function( x,
   out$family = print(x$internal$family)
   cat( "\n")
 
+  cat( "Sanity check: \n")
+  (s <- sanity(x))
+  if (!all(unlist(s))) {
+    cat("\n**Possible issues detected! Check output of sanity().**\n")
+  }
+  cat( "\n")
+
   if( !is.null(x$sdrep) ){
     cat( "\n")
     out$sdrep = print(x$sdrep)